moving remaining functions to utils file to keep main.go lighter.
